backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.1
numpy==1.26.2
scikit-learn==1.3.2
